refactor(ProductDetail): 변경된 repository 구조 적용
