+++ app/src/main/java/woowacourse/shopping/productdetail/ProductDetailPresenter.kt
@@ -2,7 +2,7 @@ package woowacourse.shopping.productdetail
 
 import model.CartProduct
 import model.Count
-import woowacourse.shopping.database.ShoppingCache
+import woowacourse.shopping.database.cart.repository.CartRepository
 import woowacourse.shopping.model.ProductUiModel
 import woowacourse.shopping.util.toProductDomainModel
 
@@ -10,7 +10,7 @@ class ProductDetailPresenter(
     product: ProductUiModel,
     private val view: ProductDetailContract.View,
     private val latestViewedProduct: ProductUiModel?,
-    private val shoppingCache: ShoppingCache,
+    private val cartRepository: CartRepository,
 ) : ProductDetailContract.Presenter {
 
     private var cartProduct: CartProduct = CartProduct(
@@ -24,7 +24,7 @@ class ProductDetailPresenter(
     }
 
     override fun addToCart() {
-        shoppingCache.insertToShoppingCart(
+        cartRepository.addToCart(
             id = cartProduct.product.id,
             count = cartProduct.count.value
         )
@@ -43,7 +43,9 @@ class ProductDetailPresenter(
     }
 
     override fun minusCartProductCount() {
-        cartProduct = cartProduct.minusCount()
-        view.setUpDialogTotalPriceView(totalPrice = cartProduct.price.value)
+        cartProduct.minusCount()?.let {
+            cartProduct = it
+            view.setUpDialogTotalPriceView(totalPrice = it.price.value)
+        }
     }
 }
