+++ app/src/main/java/woowacourse/shopping/productdetail/navigator/ProductDetailNavigatorImpl.kt
@@ -1,6 +1,7 @@
 package woowacourse.shopping.productdetail.navigator
 
 import android.content.Context
+import android.content.Intent
 import woowacourse.shopping.model.ProductUiModel
 import woowacourse.shopping.productdetail.ProductDetailActivity
 import woowacourse.shopping.shoppingcart.CartActivity
@@ -18,6 +19,8 @@ class ProductDetailNavigatorImpl(private val context: Context) : ProductDetailNa
             context = context,
             product = product,
             latestViewedProduct = null
+        ).setFlags(
+            Intent.FLAG_ACTIVITY_CLEAR_TOP
         )
 
         context.startActivity(intent)
