+++ app/src/test/java/woowacourse/shopping/ProductDetailPresenterTestUtil.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping
 
-import woowacourse.shopping.database.ShoppingCache
+import woowacourse.shopping.database.cart.repository.CartRepository
 import woowacourse.shopping.model.ProductUiModel
 import woowacourse.shopping.productdetail.ProductDetailContract
 import woowacourse.shopping.productdetail.ProductDetailPresenter
@@ -9,7 +9,7 @@ fun ProductDetailPresenter(
     product: ProductUiModel = ProductUiModel(name = "아메리카노"),
     view: ProductDetailContract.View,
     latestViewedProduct: ProductUiModel? = ProductUiModel(name = "카푸치노"),
-    repository: ShoppingCache,
+    repository: CartRepository,
 ): ProductDetailPresenter {
 
     return ProductDetailPresenter(
